requirements.txt
Pillow == 8.3.1
termcolor == 1.1.0
colorgram.py == 1.2.0